refactor: move PrecommitWithOptions to mage.go again
